requirements.txt
aiogram==3.6.0
python-dotenv==1.0.1
aiohttp==3.9.5
sqlalchemy==2.0.30
celery==5.4.0
flower==2.0.1
psycopg2-binary==2.9.9
asyncpg==0.29.0 
redis==5.0.1
aio-pika==9.4.1
pydantic_settings
requests
fastapi
uvicorn
yookassa
python-dateutil
